refactor(client-presence): split files for encapsulated support 1. Separate types where there might be multiple internal uses - especially potentially across layers
origin -> new types location
-------------------------------
broadcastControls.ts -> broadcastControlsTypes.ts
notificationsManager.ts -> notificationsManagerTypes.ts

2. Break up internalTypes.ts
   type -> new location
   -----------------------
   ExtensionRuntimeProperties, ExtensionHost, IEphemeralRuntime -> internalRuntimeTypes.ts
   ClientRecord -> stateDatastore.ts
   PostUpdateAction, ValueManager -> statesManagerTypes.ts
   Validatable* -> validatableTypes.ts
