+++ packages/framework/presence/src/latestMapValueManager.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import-x/order */
 /*!
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
@@ -12,14 +13,10 @@ import type {
 	JsonSerializable,
 } from "@fluidframework/core-interfaces/internal/exposedUtilityTypes";
 
-import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControls.js";
 import { OptionalBroadcastControl } from "./broadcastControls.js";
+import type { BroadcastControls, BroadcastControlSettings } from "./broadcastControlsTypes.js";
 import type { InternalTypes } from "./exposedInternalTypes.js";
-import type {
-	PostUpdateAction,
-	ValidatableOptionalState,
-	ValueManager,
-} from "./internalTypes.js";
+import type { ValidatableOptionalState } from "./validatableTypes.js";
 import {
 	asDeeplyReadonly,
 	asDeeplyReadonlyDeserializedJson,
@@ -40,6 +37,7 @@ import type {
 } from "./latestValueTypes.js";
 import type { AttendeeId, Attendee, Presence, SpecificAttendee } from "./presence.js";
 import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
+import type { PostUpdateAction, ValueManager } from "./statesManagerTypes.js";
 import { brandIVM } from "./valueManager.js";
 
 /**
