+++ packages/framework/presence/src/notificationsManager.ts
@@ -4,154 +4,23 @@
  */
 
 import { createEmitter } from "@fluid-internal/client-utils";
-import type { Listeners, Listenable, Off } from "@fluidframework/core-interfaces";
+import type { Listeners } from "@fluidframework/core-interfaces";
 import type { JsonTypeWith } from "@fluidframework/core-interfaces/internal";
 
 import type { InternalTypes } from "./exposedInternalTypes.js";
 import type { InternalUtilityTypes } from "./exposedUtilityTypes.js";
-import type { PostUpdateAction, ValueManager } from "./internalTypes.js";
 import { revealOpaqueJson, toOpaqueJson } from "./internalUtils.js";
+import type {
+	NotificationsManager,
+	NotificationsManagerEvents,
+	NotificationEmitter,
+	NotificationSubscriptions,
+	NotificationListenable,
+} from "./notificationsManagerTypes.js";
 import type { Attendee, PresenceWithNotifications as Presence } from "./presence.js";
 import { datastoreFromHandle, type StateDatastore } from "./stateDatastore.js";
+import type { PostUpdateAction, ValueManager } from "./statesManagerTypes.js";
 import { brandIVM } from "./valueManager.js";
-
-/**
- * @sealed
- * @alpha
- */
-export interface NotificationsManagerEvents {
-	/**
-	 * Raised when notification is received, but no subscribers were found.
-	 *
-	 * @eventProperty
-	 */
-	unattendedNotification: (name: string, sender: Attendee, ...content: unknown[]) => void;
-}
-
-/**
- * An object which allows the registration of listeners so that subscribers can be
- * notified when a notification happens.
- *
- * @sealed
- * @alpha
- */
-export interface NotificationListenable<
-	TListeners extends InternalUtilityTypes.NotificationListeners<TListeners>,
-> {
-	/**
-	 * Register a notification listener.
-	 * @param notificationName - the name of the notification
-	 * @param listener - The listener function to run when the notification is fired.
-	 * @returns A {@link @fluidframework/core-interfaces#Off | function} which will deregister the listener when called.
-	 * Calling the deregistration function more than once will have no effect.
-	 *
-	 * Listeners may also be deregistered by passing the listener to {@link NotificationListenable.off | off()}.
-	 * @remarks Registering the exact same `listener` object for the same notification more than once will throw an error.
-	 * If registering the same listener for the same notification multiple times is desired, consider using a wrapper function for the second subscription.
-	 */
-	on<K extends keyof InternalUtilityTypes.NotificationListeners<TListeners>>(
-		notificationName: K,
-		listener: (
-			sender: Attendee,
-			...args: InternalUtilityTypes.JsonDeserializedParameters<TListeners[K]>
-		) => void,
-	): Off;
-
-	/**
-	 * Deregister notification listener.
-	 * @param notificationName - The name of the notification.
-	 * @param listener - The listener function to remove from the current set of notification listeners.
-	 * @remarks If `listener` is not currently registered, this method will have no effect.
-	 *
-	 * Listeners may also be deregistered by calling the {@link @fluidframework/core-interfaces#Off | deregistration function} returned when they are {@link NotificationListenable.on | registered}.
-	 */
-	off<K extends keyof InternalUtilityTypes.NotificationListeners<TListeners>>(
-		notificationName: K,
-		listener: (
-			sender: Attendee,
-			...args: InternalUtilityTypes.JsonDeserializedParameters<TListeners[K]>
-		) => void,
-	): void;
-}
-
-/**
- * Record of notification subscriptions.
- *
- * @sealed
- * @alpha
- */
-export type NotificationSubscriptions<
-	E extends InternalUtilityTypes.NotificationListeners<E>,
-> = {
-	[K in string & keyof InternalUtilityTypes.NotificationListeners<E>]: (
-		sender: Attendee,
-		...args: InternalUtilityTypes.JsonDeserializedParameters<E[K]>
-	) => void;
-};
-
-/**
- * Interface for a notification emitter that can send typed notification to other clients.
- *
- * @sealed
- * @alpha
- */
-export interface NotificationEmitter<E extends InternalUtilityTypes.NotificationListeners<E>> {
-	/**
-	 * Emits a notification with the specified name and arguments, notifying all clients.
-	 * @param notificationName - the name of the notification to fire
-	 * @param args - the arguments sent with the notification
-	 */
-	broadcast<K extends keyof InternalUtilityTypes.NotificationListeners<E>>(
-		notificationName: K,
-		...args: Parameters<E[K]>
-	): void;
-
-	/**
-	 * Emits a notification with the specified name and arguments, notifying a single attendee.
-	 * @param notificationName - the name of the notification to fire
-	 * @param targetAttendee - the single attendee to notify
-	 * @param args - the arguments sent with the notification
-	 */
-	unicast<K extends keyof InternalUtilityTypes.NotificationListeners<E>>(
-		notificationName: K,
-		targetAttendee: Attendee,
-		...args: Parameters<E[K]>
-	): void;
-}
-
-/**
- * Provides notifications from this client to others and subscription
- * to their notifications.
- *
- * @remarks Create using {@link Notifications} registered to
- * {@link NotificationsWorkspace} or {@link StatesWorkspace}.
- *
- * @sealed
- * @alpha
- */
-export interface NotificationsManager<
-	T extends InternalUtilityTypes.NotificationListeners<T>,
-> {
-	/**
-	 * Containing {@link Presence}
-	 */
-	readonly presence: Presence;
-
-	/**
-	 * Events for Notifications manager.
-	 */
-	readonly events: Listenable<NotificationsManagerEvents>;
-
-	/**
-	 * Send notifications to other clients.
-	 */
-	readonly emit: NotificationEmitter<T>;
-
-	/**
-	 * Provides subscription to notifications from other clients.
-	 */
-	readonly notifications: NotificationListenable<T>;
-}
 
 /**
  * Object.keys retyped to support specific records keys and
