+++ packages/framework/presence/src/presenceDatastoreManager.ts
@@ -9,15 +9,9 @@ import { assert } from "@fluidframework/core-utils/internal";
 import type { ITelemetryLoggerExt } from "@fluidframework/telemetry-utils/internal";
 
 import type { ClientConnectionId } from "./baseTypes.js";
-import type { BroadcastControlSettings } from "./broadcastControls.js";
+import type { BroadcastControlSettings } from "./broadcastControlsTypes.js";
 import type { InternalTypes } from "./exposedInternalTypes.js";
-import type {
-	IEphemeralRuntime,
-	PostUpdateAction,
-	ValidatableOptionalState,
-	ValidatableValueDirectory,
-	ValidatableValueStructure,
-} from "./internalTypes.js";
+import type { IEphemeralRuntime } from "./internalRuntimeTypes.js";
 import { objectEntries } from "./internalUtils.js";
 import type {
 	AttendeeId,
@@ -50,6 +44,7 @@ import {
 	datastoreUpdateMessageType,
 	joinMessageType,
 } from "./protocol.js";
+import type { PostUpdateAction } from "./statesManagerTypes.js";
 import type { SystemWorkspaceDatastore } from "./systemWorkspace.js";
 import { TimerManager } from "./timerManager.js";
 import type {
@@ -60,6 +55,11 @@ import type {
 	StatesWorkspaceSchema,
 	WorkspaceAddress,
 } from "./types.js";
+import type {
+	ValidatableOptionalState,
+	ValidatableValueDirectory,
+	ValidatableValueStructure,
+} from "./validatableTypes.js";
 
 interface AnyWorkspaceEntry<TSchema extends StatesWorkspaceSchema> {
 	public: AnyWorkspace<TSchema>;
