+++ packages/framework/presence/src/types.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT License.
  */
 
-import type { BroadcastControls } from "./broadcastControls.js";
+import type { BroadcastControls } from "./broadcastControlsTypes.js";
 import type { InternalTypes } from "./exposedInternalTypes.js";
-import type { NotificationsManager } from "./notificationsManager.js";
+import type { NotificationsManager } from "./notificationsManagerTypes.js";
 import type { Presence, PresenceWithNotifications } from "./presence.js";
 
 /**
